+++ infrastructure-android/src/main/kotlin/android/AndroidConfigPlugin.kt
@@ -1,6 +1,7 @@
 package com.redmadrobot.build.android
 
 import com.redmadrobot.build.InfrastructurePlugin
+import com.redmadrobot.build.StaticAnalyzerSpec
 import com.redmadrobot.build.kotlin.KotlinConfigPlugin
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
@@ -20,6 +21,9 @@ public open class AndroidConfigPlugin : InfrastructurePlugin() {
 
     internal val jvmTarget: Property<JavaVersion>
         get() = redmadrobotExtension.jvmTarget
+
+    internal val staticAnalyzerSpec: StaticAnalyzerSpec
+        get() = redmadrobotExtension
 
     override fun Project.configure() {
         val kotlinConfigPlugin = plugins.apply(KotlinConfigPlugin::class)