+++ infrastructure-android/src/main/kotlin/android/internal/Accessors.kt
@@ -1,15 +1,14 @@
 package com.redmadrobot.build.android.internal
 
-import com.android.build.gradle.BaseExtension
+import com.android.build.api.dsl.CommonExtension
 import com.redmadrobot.build.android.AndroidOptions
 import com.redmadrobot.build.kotlin.TestOptions
 import org.gradle.api.Project
 import org.gradle.api.plugins.ExtensionAware
 import org.gradle.kotlin.dsl.getByName
 
-internal val Project.android: BaseExtension get() = extensions.getByName<BaseExtension>("android")
-
-internal fun <T : BaseExtension> Project.android(configure: T.() -> Unit) {
+@Suppress("UnstableApiUsage")
+internal fun <T : CommonExtension<*, *, *, *>> Project.android(configure: T.() -> Unit) {
     extensions.configure("android", configure)
 }